--- bpf/tests/tc_nodeport_lb4_fragments_ns.c
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: (GPL-2.0-only OR BSD-2-Clause)
-/* Copyright Authors of Cilium */
-
-#define NORTH_SOUTH_TEST 1
-
-#include "tc_nodeport_lb4_fragments.h"
