+++ bpf/tests/tc_nodeport_lb_fragments.h
@@ -6,17 +6,20 @@
 #include "common.h"
 #include "pktgen.h"
 
-#define ENABLE_IPV4 1
-#define ENABLE_NODEPORT 1
+#define ENABLE_IPV4		1
+#define ENABLE_IPV6		1
+#define ENABLE_NODEPORT		1
 
 #ifdef NORTH_SOUTH_TEST
 # define CLIENT_IP	v4_ext_one
+# define CLIENT_IPV6	v6_ext_node_one
 # define HOOK		netdev_receive_packet
 
 # include "lib/bpf_host.h"
 #elif defined(EAST_WEST_TEST)
 # define ENABLE_SOCKET_LB_HOST_ONLY 1
 # define CLIENT_IP	v4_pod_two
+# define CLIENT_IPV6	v6_pod_two
 # define HOOK		pod_send_packet
 
 # include "lib/bpf_lxc.h"
@@ -26,6 +29,7 @@
 
 ASSIGN_CONFIG(bool, enable_conntrack_accounting, true)
 ASSIGN_CONFIG(bool, enable_ipv4_fragments, true)
+ASSIGN_CONFIG(bool, enable_ipv6_fragments, true)
 
 #include "lib/endpoint.h"
 #include "lib/policy.h"
@@ -35,9 +39,11 @@ ASSIGN_CONFIG(bool, enable_ipv4_fragments, true)
 #define CLIENT_PORT	tcp_src_one
 
 #define FRONTEND_IP	v4_svc_one
+#define FRONTEND_IPV6	v6_svc_one
 #define FRONTEND_PORT	tcp_svc_one
 
 #define BACKEND_IP	v4_pod_one
+#define BACKEND_IPV6	v6_pod_one
 #define BACKEND_PORT	tcp_dst_one
 
 #define NAT_REV_INDEX	1
@@ -106,6 +112,62 @@ const __u8 lb4_ew_nodeport_fragment1_post_dnat[] = {
 /* packet defined in ./scapy/lb_pkt_defs.py */
 const __u8 lb4_ew_nodeport_fragment2_post_dnat[] = {
 	SCAPY_BUF_BYTES(lb4_ew_nodeport_fragment2_post_dnat)
+};
+
+/* Expected CT tuple to check for conntrack */
+#define EXPECTED_CT_TUPLE6 {\
+	.saddr = *(union v6addr *)CLIENT_IPV6, \
+	.sport = FRONTEND_PORT, \
+	.daddr = *(union v6addr *)FRONTEND_IPV6, \
+	.dport = CLIENT_PORT, \
+	.nexthdr = IPPROTO_TCP, \
+	.flags = TUPLE_F_SERVICE }
+
+/* Expected fragment ID to check for conntrack */
+#define EXPECTED_FRAG_ID6 { \
+	.saddr = *(union v6addr *)CLIENT_IPV6, \
+	.daddr = *(union v6addr *)FRONTEND_IPV6, \
+	.id = bpf_htonl(256), \
+	.proto = IPPROTO_TCP }
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ns_nodeport_fragment1[] = {
+	SCAPY_BUF_BYTES(lb6_ns_nodeport_fragment1)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ns_nodeport_fragment2[] = {
+	SCAPY_BUF_BYTES(lb6_ns_nodeport_fragment2)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ns_nodeport_fragment1_post_dnat[] = {
+	SCAPY_BUF_BYTES(lb6_ns_nodeport_fragment1_post_dnat)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ns_nodeport_fragment2_post_dnat[] = {
+	SCAPY_BUF_BYTES(lb6_ns_nodeport_fragment2_post_dnat)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ew_nodeport_fragment1[] = {
+	SCAPY_BUF_BYTES(lb6_ew_nodeport_fragment1)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ew_nodeport_fragment2[] = {
+	SCAPY_BUF_BYTES(lb6_ew_nodeport_fragment2)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ew_nodeport_fragment1_post_dnat[] = {
+	SCAPY_BUF_BYTES(lb6_ew_nodeport_fragment1_post_dnat)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ew_nodeport_fragment2_post_dnat[] = {
+	SCAPY_BUF_BYTES(lb6_ew_nodeport_fragment2_post_dnat)
 };
 
 /* Test that the 1st fragment of an external-to-nodeport request is handled correctly */
@@ -271,6 +333,172 @@ int nodeport_lb4_fragments_2_check(struct __ctx_buff *ctx)
 
 	/* Ensure CT entry is updated accordingly (SVC). */
 	ct_entry = map_lookup_elem(get_ct_map4(&expected_ct_tuple), &expected_ct_tuple);
+	assert(ct_entry);
+	assert(ct_entry->packets == count);
+	assert(ct_entry->bytes == bytes);
+
+	test_finish();
+}
+
+/* Test that the 1st fragment is handled correctly */
+PKTGEN("tc", "tc_nodeport_lb6_fragment1")
+int nodeport_lb6_fragment1_pktgen(struct __ctx_buff *ctx)
+{
+	struct pktgen builder;
+
+	pktgen__init(&builder, ctx);
+
+#ifdef NORTH_SOUTH_TEST
+	scapy_push_data(&builder, lb6_ns_nodeport_fragment1, sizeof(lb6_ns_nodeport_fragment1));
+#else
+	scapy_push_data(&builder, lb6_ew_nodeport_fragment1, sizeof(lb6_ew_nodeport_fragment1));
+#endif
+
+	pktgen__finish(&builder);
+
+	return 0;
+}
+
+SETUP("tc", "tc_nodeport_lb6_fragment1")
+int nodeport_lb6_fragment1_setup(struct __ctx_buff *ctx)
+{
+	policy_add_egress_allow_all_entry();
+
+	endpoint_v6_add_entry((union v6addr *)BACKEND_IPV6, BACKEND_IFINDEX, 0, 0, 0,
+			      (__u8 *)mac_one, (__u8 *)mac_two);
+	lb_v6_add_service_with_flags((union v6addr *)FRONTEND_IPV6, FRONTEND_PORT, IPPROTO_TCP,
+				     BACKEND_COUNT, NAT_REV_INDEX,
+				     SVC_FLAG_ROUTABLE | SVC_FLAG_NODEPORT, 0);
+	lb_v6_add_backend((union v6addr *)FRONTEND_IPV6, FRONTEND_PORT, BACKEND_COUNT, BACKEND_ID,
+			  (union v6addr *)BACKEND_IPV6, BACKEND_PORT, IPPROTO_TCP, 0);
+
+	return HOOK(ctx);
+}
+
+CHECK("tc", "tc_nodeport_lb6_fragment1")
+int nodeport_lb6_fragment1_check(struct __ctx_buff *ctx)
+{
+	__u32 *status_code;
+	void *data_end;
+	void *data;
+	struct ipv6_frag_l4ports *l4ports;
+	struct ct_entry *ct_entry;
+	struct ipv6_frag_id expected_frag_id = EXPECTED_FRAG_ID6;
+	struct ipv6_ct_tuple expected_ct_tuple = EXPECTED_CT_TUPLE6;
+	struct metrics_key metric_key = EXPECTED_METRIC_KEY;
+	__u64 metric_count = 3;
+	__u64 count = 1;
+	__u64 bytes = 0;
+
+	test_init();
+
+	data = ctx_data(ctx);
+	data_end = ctx_data_end(ctx);
+	status_code = data;
+
+	/* Ensure packet offset and status code. */
+	assert(data + sizeof(__u32) <= data_end);
+	assert(*status_code == CTX_ACT_OK);
+
+	/* Ensure fragment L4 ports are being tracked. */
+	l4ports = map_lookup_elem(&cilium_ipv6_frag_datagrams, &expected_frag_id);
+	assert(l4ports);
+	assert(l4ports->dport == FRONTEND_PORT);
+	assert(l4ports->sport == CLIENT_PORT);
+
+	/* Ensure fragment-related metrics are updated accordingly. */
+	assert_metrics_count(metric_key, metric_count);
+
+	/* Ensure packet has been DNAT correctly. */
+#ifdef NORTH_SOUTH_TEST
+	ASSERT_CTX_BUF_OFF("tcp6_first_fragment_ok", "Ether", ctx, sizeof(__u32),
+			   lb6_ns_nodeport_fragment1_post_dnat,
+			   sizeof(lb6_ns_nodeport_fragment1_post_dnat));
+	bytes = sizeof(lb6_ns_nodeport_fragment1_post_dnat);
+#else
+	ASSERT_CTX_BUF_OFF("tcp6_first_fragment_ok", "Ether", ctx, sizeof(__u32),
+			   lb6_ew_nodeport_fragment1_post_dnat,
+			   sizeof(lb6_ew_nodeport_fragment1_post_dnat));
+	bytes = sizeof(lb6_ew_nodeport_fragment1_post_dnat);
+#endif
+
+	/* Ensure CT entry is updated accordingly (SVC). */
+	ct_entry = map_lookup_elem(get_ct_map6(&expected_ct_tuple), &expected_ct_tuple);
+	assert(ct_entry);
+	assert(ct_entry->packets == count);
+	assert(ct_entry->bytes == bytes);
+
+	test_finish();
+}
+
+/* Test that the 2nd fragment is handled correctly */
+PKTGEN("tc", "tc_nodeport_lb6_fragment2")
+int nodeport_lb6_fragment2_pktgen(struct __ctx_buff *ctx)
+{
+	struct pktgen builder;
+
+	pktgen__init(&builder, ctx);
+
+#ifdef NORTH_SOUTH_TEST
+	scapy_push_data(&builder, lb6_ns_nodeport_fragment2, sizeof(lb6_ns_nodeport_fragment2));
+#else
+	scapy_push_data(&builder, lb6_ew_nodeport_fragment2, sizeof(lb6_ew_nodeport_fragment2));
+#endif
+
+	pktgen__finish(&builder);
+
+	return 0;
+}
+
+SETUP("tc", "tc_nodeport_lb6_fragment2")
+int nodeport_lb6_fragment2_setup(struct __ctx_buff *ctx)
+{
+	return HOOK(ctx);
+}
+
+CHECK("tc", "tc_nodeport_lb6_fragment2")
+int nodeport_lb6_fragment2_check(struct __ctx_buff *ctx)
+{
+	__u32 *status_code;
+	void *data_end;
+	void *data;
+	struct ct_entry *ct_entry;
+	struct ipv6_ct_tuple expected_ct_tuple = EXPECTED_CT_TUPLE6;
+	struct metrics_key metric_key = EXPECTED_METRIC_KEY;
+	__u64 count = 2;
+	__u64 metric_count = 4;
+	__u64 bytes = 0;
+
+	test_init();
+
+	data = ctx_data(ctx);
+	data_end = ctx_data_end(ctx);
+	status_code = data;
+
+	/* Ensure packet offset and status code. */
+	assert(data + sizeof(__u32) <= data_end);
+	assert(*status_code == CTX_ACT_OK);
+
+	/* Ensure fragment-related metrics are updated accordingly. */
+	assert_metrics_count(metric_key, metric_count);
+
+	/* Ensure packet has been DNAT correctly. */
+#ifdef NORTH_SOUTH_TEST
+	ASSERT_CTX_BUF_OFF("tcp6_second_fragment_ok", "Ether", ctx, sizeof(__u32),
+			   lb6_ns_nodeport_fragment2_post_dnat,
+			   sizeof(lb6_ns_nodeport_fragment2_post_dnat));
+	bytes = sizeof(lb6_ns_nodeport_fragment1_post_dnat) +
+		sizeof(lb6_ns_nodeport_fragment2_post_dnat);
+#else
+	ASSERT_CTX_BUF_OFF("tcp6_second_fragment_ok", "Ether", ctx, sizeof(__u32),
+			   lb6_ew_nodeport_fragment2_post_dnat,
+			   sizeof(lb6_ew_nodeport_fragment2_post_dnat));
+	bytes = sizeof(lb6_ew_nodeport_fragment1_post_dnat) +
+		sizeof(lb6_ew_nodeport_fragment2_post_dnat);
+#endif
+
+	/* Ensure CT entry is updated accordingly (SVC). */
+	ct_entry = map_lookup_elem(get_ct_map6(&expected_ct_tuple), &expected_ct_tuple);
 	assert(ct_entry);
 	assert(ct_entry->packets == count);
 	assert(ct_entry->bytes == bytes);
