+++ bpf/tests/tc_nodeport_lb_fragments_ns.c
@@ -3,4 +3,4 @@
 
 #define NORTH_SOUTH_TEST 1
 
-#include "tc_nodeport_lb6_fragments.h"
+#include "tc_nodeport_lb_fragments.h"
